I want to throw myself over a bridge
Okay so, our project wasn't training for 29 days and no one noticed? Why the hell did we build all this production pipeline and testing automation if we can't even detect when we break down %50 of the project? Only reason i found this is because i wanted to check if i could directly port the new class for markopy cli. How does something like this happen? I mean i understand that i was tipsy when making some of the code changes but what the hell people, does no one ever test our shit?